Update entities and script to correctly parse date
