+++ backend/entities/academic_advising/drop_in_entity.py
@@ -22,6 +22,8 @@ class DropInEntity(EntityBase):
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     # Title of the drop-in session (whose session this is)
     title: Mapped[str] = mapped_column(String)
+    # Date of the drop-in session
+    date: Mapped[datetime] = mapped_column(DateTime)
     # Start time of the drop-in session
     start: Mapped[datetime] = mapped_column(DateTime)
     # End time of the drop-in session
@@ -50,6 +52,7 @@ class DropInEntity(EntityBase):
         return cls(
             id=model.id,
             title=model.title,
+            date=model.date,
             start=model.start,
             end=model.end,
             link=model.link,
@@ -60,6 +63,7 @@ class DropInEntity(EntityBase):
         return DropIn(
             id=self.id,
             title=self.title,
+            date=self.date,
             start=self.start,
             end=self.end,
             link=self.link,
